Move shops to shared module
