+++ src/api/main/scala/training/modules/shops/CommercialActivityRepository.scala
@@ -2,6 +2,7 @@ package training.modules.shops
 
 import doobie.ConnectionIO
 import doobie.implicits.toSqlInterpolator
+import training.domain.CommercialActivity
 
 class CommercialActivityRepository {
