+++ src/api/main/scala/training/modules/shops/StratumRepository.scala
@@ -2,6 +2,7 @@ package training.modules.shops
 
 import doobie.ConnectionIO
 import doobie.implicits.toSqlInterpolator
+import training.domain.Stratum
 
 class StratumRepository {
   def find(stratumId: Int): ConnectionIO[Option[Stratum]] =
