+++ src/scrapper/main/scala/training/scrapper/modules/apiclient/ApiClient.scala
@@ -1,0 +1,7 @@
+package training.scrapper.modules.apiclient
+
+import training.domain.Shop
+
+trait ApiClient {
+  def call(shops: Seq[Shop]): Either[Error, Unit]
+}
