+++ src/scrapper/main/scala/training/scrapper/modules/shared/Error.scala
@@ -1,0 +1,3 @@
+package training.scrapper.modules.shared
+
+case class Error(message: String, e: Option[Exception] = None)
