+++ src/scrapper/main/scala/training/scrapper/modules/webcollector/WebCollector.scala
@@ -1,0 +1,8 @@
+package training.scrapper.modules.webcollector
+
+import training.domain.Shop
+import training.scrapper.modules.shared.Error
+
+trait WebCollector {
+  def invoke(url: String): Either[Error, Seq[Shop]]
+}
