+++ src/shared/test/scala/training/BaseSpec.scala
